+++ spring-cloud-gcp-data-spanner/pom.xml
@@ -44,7 +44,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-gcp-test-support</artifactId>
+			<artifactId>spring-cloud-gcp-test</artifactId>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
