+++ spring-cloud-gcp-dependencies/pom.xml
@@ -192,7 +192,7 @@
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
-				<artifactId>spring-cloud-gcp-test-support</artifactId>
+				<artifactId>spring-cloud-gcp-test</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<!-- spring-cloud-gcp-starter-sql -->
